Added a "property chain" to conditions of a query.
This is not the the creation of DM::Query::Path. It is the ability
to specify a path as follows:

  Zoo.all('Zoo.animals.name' => 'Monkey')

The key must be given as a string in dot (.) notation. Any table
that is not linked or included in through :links and :includes option
is automatically added to the :links option.

With this ability, do we need DM::Query::Path ??