Added Model#field_naming_conventions
As a cleaner way of handling legacy databases, where the field naming conventions can be different from one model to the next, the naming conventions for fields are now, like Model#storage_names, tied to a particular model and repository. The default convention is set to repository(k).adapter.field_naming_convention, as it was before.