Fix bugs \#444 and \#566 and \#577 which all had the same cause
This basically tries to make relationship initalization a bit smarter.
The code is pretty ugly, but it just tries loading up relations until
they work (that is, all related objects are loaded). This is done
for every Model#properties call, but of course cached once they are
in a sane state.