Ensure the correct ordering of links for more complex joins
Consider a query like the following:

Person.all(Person.memberships.group.type => 'cool', Person.memberships.valid_from.gte => Date.today)

This can result in one of the following to link lists being set up (depending on non-deterministic
hash ordering).

[:memberships, :group, :memberships]
[:group, :memberships, :memberships]

Query#normalize_links walks through this list and creates the following structure:

[:memberships, :group] or [:group, :memberships]

If the first situation happens and the subsequent join statement is created, it will
create an invalid join statement. DataObjectsAdapter#join_statement walks through
the normalized list in revere order, so if that adds group to the query before
memberships in a join, the query will fail.

The fix is very simple. Because of how the initial list is constructed, a correctly
normalized list will be created by walking the initial 3 element list given above in the
reverse order. That will always result in the [:group, :memberships] links that
will work correctly if joined in reverse order