OrOperation should be valid if at least one branch is valid
OrOperations are a bit different from the other operations, since
they are valid if at least one of the branches it contains is valid.

A future refactor should also ensure that only the valid operations
go to the datastore, this can also eliminate OR branches of only
one valid branch is left.