Added Gemfile.lock to .gitignore (that's OK for gems)
While applications normally put the Gemfile.lock
into version control, it seems unnecessary when
using the Gemfile for setting up a development
environment for a gem (like dm-core).

Using bundle exec in a locked bundle runs much
faster than in an unlocked bundle, which is
convenient when running specs often. The only
thing to take care of is that instead of just
running "bundle install" to install and update
the bundle, you should do

  bundle install
  bundle lock
  bundle install --relock # to update

Since we always want our Gemfile to point to latest
master branches, it makes no sense for us to share
the Gemfile.lock in git like applications would do
to ensure consistent deploy states. Instead, we
add it to .gitignore to make sure we never lock
the main repository by accident.