The spec/lib/adapter_helpers don't really require dm-migrations
All the calls to auto_migrate!/upgrade! are properly
shielded with respond_to? checks so strictly speaking
we aren't requiring dm-migrations to be present at
this stage. We do require dm-migrations at an earlier
point now tho, because most of the adapters typically
used for testing require an automigration procedure
to happen.