Use DataMapper::Spec.setup for setting up spec runs
Support for the ADAPTERS env variable, and thus for
running specs on multiple adapters with one command,
has been dropped with this commit. Running specs is
now limited to a single adapter at a time. Time
has shown that this is the only reasonable way to
run the specs during development. Possible CI tools
can of course still maintain a list of available
adapters and run specs for each of those separately.

Additionally, this commit forms the base of being
able to run dm-core specs on any of the available
adapters, using any of the available plugins, simply
by switching the ADAPTER and PLUGINS env variables.

Effectively, dm-core's spec_helper.rb now does

  require 'dm-core/spec/setup'
  DataMapper::Spec.setup

and expects that to properly setup the connection.
Once done, the current adapter is available via

  DataMapper::Spec.adapter

To achieve that, dm-core now provides a base spec
adapter class, located in dm-core/spec/setup.rb,
that makes it very easy to write a

  dm-*-adapter/spec/setup.rb

file for any third party provided DM adapter.

Adapter authors should have a look at

  dm-core/lib/dm-core/spec/setup

and at the various dm-*-adapter/spec/setup.rb files
below the datamapper github account, to get an idea
for how to write such a file.

Third party adapter authors are encouraged to
update their adapters and provide a

  dm-*-adapter/spec/setup.rb

file. Providing this file makes it possible for
dm-core to run its full specsuite on the adapter,
which is a quite solid test for any adapter out
there (and of course dm-core itself). The way it's
done is that basically, dm-core issues a

  require "dm-#{ENV['ADAPTER']}-adapter/spec/setup"

during the course of DataMapper::Spec.setup.

Furthermore, this commit forms the base for
simplifying spec_helper.rb files among all of the
dm-* gems. Again, all that needs to be done to
run any plugin's specs against any of the available
adapters, is to add

  require 'dm-core/spec/setup'
  DataMapper::Spec.setup

to the plugin's spec_helper.rb file. This will
make sure that the desired adapter is setup, and
the specs are ready to go. As with dm-core itself,
the current adapter is available via

  DataMapper::Spec.adapter

and of course the PLUGINS env variable is recognized
too, so if plugin specs should run using other
plugins enabled, simply run the spec command with
something like

  PLUGINS=dm-constraints,... bundle exec rake spec

and DataMapper::Spec.setup will require the plugins
for you.

Last but not least, to provide a uniform (database)
naming scheme for spec runs, we decided that we will
default to

  username: datamapper
  password: datamapper
  database: datamapper_default_tests
  database: datamapper_alternate_tests

Make sure you have these databases setup with the
appropriate privileges on every adapter that you
plan to run dm-core specs on. We need two databases
because datamapper allows to work with two of them
at a time and contains specs for doing that.