DataMapper::Spec.setup now performs adapter setup
The following methods on DataMapper::Spec now
delegate directly to the instance methods on the
currently used spec adapter.

  DataMapper::Spec.setup    # setup and memoize
  DataMapper::Spec.setup!   # always call setup
  DataMapper::Spec.adapter  # same as setup

If the ADAPTER and PLUGINS haven't been configured
at the time one of these methods is invoked, it
will be configured before proceeding with setup.

This cuts down the number of code necessary to get
specs with any adapter and plugin combination up
and running. All that needs to be done is

  require 'dm-core/spec/setup'
  DataMapper::Spec.setup # or .adapter

at the appropriate time, in specs or standalone
snippets.