Added DataMapper::Adapters::AbstractAdapter.supports_transactions?
This is a semipublic api addition that is at least
used in dm-transactions for now. It returns false
in the case of AbstractAdapter and is overwritten
to return true in DataObjectsAdapter.

This will make it easier to discover if it is
necessary to include transaction related behavior
in case dm-transactions is loaded and a new adapter
gets set up.

Also, this allows other plugins to easily discover
if they need to take special measures in case the
adapter in use supports transactions.