Don't try to do collection matching on using the related collection
Trying to do this match will potentially load up a large collection
if the referred collection is not loaded yet. This unnecessary loading
can make it really slow using these objects.

The spec change is needed because this also exposed a somewhat unrelated
issue because now the existing resource is reused which has some lazy fields
loaded. Due to the mismatch in loaded fields, the spec failed. I changed the
spec to only test the actually attribute that was changed before the save
to verify that it does work.