Remove the need for creating a Gemfile.local
This commit allows the use of local sources in the
bundle without actually creating a Gemfile.local.

Instead of running `rake local_gemfile`, you can
now just set the following env var for any of the
bundler tasks

  SOURCE=path

If this is set during the execution of any bundler
command, bundler will use local sources instead of
the :git sources which are the default.

The other environment variables (PLUGINS, ADAPTERS
and DO_GIT) are still available and haven't changed
their meaning.

If you want to use local sources instead of :git,
all you need to do is pass SOURCE=path.

WARNING:
This commit currently breaks dm-dev. As soon as all
the datamapper gemfiles are changed to behave like
this, dm-dev will be adapted to make use of this
new style.