Always access relationships in order of definition
This refactoring introduces a few new set classes
that are used to store relationships, properties
and ultimately also descendant sets.

  # Implements set behavior and
  # keeps track of insertion order
  DataMapper::OrderedSet

  # Uses an OrderedSet to keep track of
  # objects that must respond_to?(:name)
  DataMapper::SubjectSet "has a" OrderedSet

SubjectSet forms the base class for the already
present DataMapper::PropertySet and the newly added
DataMapper::RelationshipSet.

  PropertySet     < SubjectSet
  RelationshipSet < SubjectSet

This makes sure that common behavior between these
two classes is shared, and that RelationshipSet now
remembers the order in which relationships were added
too. Previously, relationships were stored inside a
hash, which would loose order on any not ruby-1.9
implementation.

By using an ordered set to store the relationships,
we can now provide consistent foreign key property
order when automigrating models (or fetching records
using DataMapper::Model#get).

This means that it's finally save to always omit an
explicit property declaration when defining any
relationship either with #belongs_to or #has.
Previously, it wasn't save to call Model#get on
a model with a CPK that consists of FK properties
(a join table). This was because Model#get relies
on the key property order which was undefined
before this patch because it simply used a Hash
internally.

[#1398 state:resolved]