Make Property lookups defensive against same-named custom properties.
Background: Property class lookups are triggered through const_missing,
typically when defining a property in the Model DSL and referring to it by
innermost class name (e.g. "Serial").  This in turn maps to Property#find_class,
which uses an algorithm that effectively relegates lookups inside the
descendants tree to a flat namespace of demodulized names.  The algorithm
returns the most-recently-defined property of the same name, which is incorrect
in the case where a custom property with the same class name as a DM built-in is
present, but not relevant in all contexts. (FYI this problem was present before
the demodulized_names optimization was introduced.)

We solve this problem by only ever adding to the lookup table (not allowing
overrides).  Given that DM typically gets its property classes loaded first,
this means its types are always "reserved", and anything externally-derived will
be included for indirect lookups only when they don't override built-in type
names.  For the common case, this should be fine.

For external property authors who want to provide "replacements" for builtins
(e.g. in a non-DM-supported adapter), they should follow the convention of
wrapping those (or all) properties in a module, and include'ing the module on
the model class directly.  This bypasses the const_missing lookup that would
normally check the Property descendant tree/table, ensuring the correct property
class mapping in the context of a given model.