Load model keys when setting up a resource
This makes sure that in case of using custom properties as a model key
correct values will be used to store a resource in the Identity Map.
Otherwise there is a discrepancy between the way resources are stored
and retrieved from IM which results in a resource in the Immutable
state (and some other weird side-effects).

This was a known issue for years but in the beginning there was a theory
that the adapter should return values already loaded. Now situation is
different because load/dump in property are not related to datastore
capabilities, they are used to change the way how properties are stored and
retrieved from the datastore. For example Enum is dumped to an integer and
loaded to a string or UUID is dumped to a string but loaded to UUIDTools::UUID
instance etc.

[#1478 state:resolved]