Refactor properties and relationships to use a single module
This prevents having a very large number of anonymous modules
attached to a model. All these anonymous models harm performance
because of slower method lookup. It also makes the behavior when
reopening a model and redefining a property more consistent. The
latter actually happens a lot in the specs.

This was found wheni debugging the JIT issue with Rubinius.